requirements.txt
fastapi==0.68.1
python-multipart==0.0.5
pytesseract==0.3.8
pdf2image==1.16.0
spacy==3.1.3
python-docx==0.8.11
PyMuPDF==1.19.1
uvicorn==0.15.0
python-jose==3.3.0 